# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "One Medy API Documentation",
        "description": "One Medy APi Service to be consumed by client Application",
        "contact": {
            "email": "info@walexbiznig.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/labtest/add": {
            "post": {
                "tags": [
                    "Labtest"
                ],
                "summary": "Create a new lab test for a user.",
                "operationId": "1c6f7157e4c06ec4fc68b0e43c078979",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLabtestRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/user/labtest/{id}": {
            "get": {
                "tags": [
                    "Labtest"
                ],
                "summary": "Returns a user lab tests",
                "description": "Returns lab tests of provided user id",
                "operationId": "id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User Id",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/doctor/labtest/{id}": {
            "get": {
                "tags": [
                    "Labtest"
                ],
                "summary": "Return lab tests from a doctor",
                "description": "Returns lab tests made by the provided doctor id",
                "operationId": "id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Doctor Id",
                        "required": true
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/lab/external/{id}": {
            "patch": {
                "tags": [
                    "Labtest"
                ],
                "summary": "change the state of isexternallab column to 1",
                "operationId": "8765dc4d6211dcaeef6c7df29095d6ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment Id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIsExternalLabRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/lab/internal/{id}": {
            "patch": {
                "tags": [
                    "Labtest"
                ],
                "summary": "change the state of isexternallab column to 0",
                "operationId": "be5f2fa2e0e4b3581276ebb53b357a92",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment Id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIsExternalLabRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/labtests": {
            "get": {
                "tags": [
                    "Labtest"
                ],
                "summary": "Return a json array of all the labtests sent to the user.",
                "description": "Returns all lab tests both the completed and uncompleted. If isdoctorended is placed in the URL as parameter with a value of\r\n     * 1, it will return all completed tests. The opposite happens if the value of isdoctorended is made 0",
                "operationId": "getAllLabtests",
                "parameters": [
                    {
                        "name": "isdoctorended",
                        "in": "parameter",
                        "description": "Status of the lab test",
                        "required": false
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/isdoctorended/{id}": {
            "patch": {
                "tags": [
                    "Labtest"
                ],
                "summary": "change the state of isdoctorended column to 1 and changes the appointment step to 3",
                "operationId": "30f8f9fd5f4e2f62fdfb5b6d2c16094b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment Id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIsDoctorEndedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/v1/updateresult/labtest/{id}": {
            "patch": {
                "tags": [
                    "Labtest"
                ],
                "summary": "Update the result in the labtest table",
                "operationId": "4a5250d3f95b80a14c14c667bfe15313",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment Id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResultRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateLabtestRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "title": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "appointment_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "result": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "properties": {
                            "description": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/CreateLabtestRequest"
                    }
                ]
            },
            "UpdateIsDoctorEndedRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "appointment_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/UpdateIsDoctorEndedRequest"
                    }
                ]
            },
            "UpdateIsExternalLabRequest": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "appointment_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/UpdateIsExternalLabRequest"
                    }
                ]
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}